Build: Fixing unit tests, now with pipe()
As subprocess.Popen needs a file descriptor on
the "file-like object" passed as stderr and/or
stdout, I'm using the os.pipe call to create
a pipe. Then I write to the write side of the pipe
and close it. This causes the nice effect of us
having a file like object with a file descriptor.